Works with Mizi heat pump

+++ Makefile
@@ -1,7 +1,7 @@
 push:
-	rm *.bin
+	[ -f *.bin ] && rm *.bin
 	particle compile photon
-	particle flash bager_bager *bin
+	particle flash bager_bager *.bin
 
 clean:
 	rm *.bin
